Revert "ExtensionResourceParser"
This reverts commit 2b5cbebd9de3d4d0e1ee26ac21224b477c1b7183.
