+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/parsers/v2/ExtensionResourceParser.java
@@ -7,24 +7,18 @@ import java.util.Map;
 import com.sap.cloud.lm.sl.common.ParsingException;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionResource;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionResource.Builder;
-import com.sap.cloud.lm.sl.mta.parsers.ModelParser;
 import com.sap.cloud.lm.sl.mta.schema.MapElement;
 
-public class ExtensionResourceParser extends ModelParser<ExtensionResource> {
+public class ExtensionResourceParser extends com.sap.cloud.lm.sl.mta.parsers.v1.ExtensionResourceParser {
 
-    protected static final String PROCESSED_OBJECT_NAME = "MTA extension resource";
-
-    public static final String NAME = "name";
-    public static final String TYPE = "type";
-    public static final String PROPERTIES = "properties";
     public static final String PARAMETERS = "parameters";
 
     public ExtensionResourceParser(Map<String, Object> source) {
-        this(EXT_RESOURCE, source);
+        super(EXT_RESOURCE, source);
     }
 
     protected ExtensionResourceParser(MapElement schema, Map<String, Object> source) {
-        super(PROCESSED_OBJECT_NAME, schema, source);
+        super(schema, source);
     }
 
     @Override
@@ -34,14 +28,6 @@ public class ExtensionResourceParser extends ModelParser<ExtensionResource> {
         builder.setParameters(getParameters());
         builder.setProperties(getProperties());
         return builder.build();
-    }
-
-    protected String getName() {
-        return getStringElement(NAME);
-    }
-
-    protected Map<String, Object> getProperties() {
-        return getMapElement(PROPERTIES);
     }
 
     protected Map<String, Object> getParameters() {
